% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{sq_api_token}
\alias{sq_api_token}
\title{Get or set your API token}
\usage{
sq_api_token(token)
}
\arguments{
\item{token}{string: token. If not provided, and the `SQUIDLE_API_TOKEN` environment variable is set (e.g. via `Sys.setenv(SQUIDLE_API_TOKEN = "xyz")`) that will be used}
}
\value{
The API token (invisibly unless the `token` parameter was not provided)
}
\description{
Get or set your API token
}
\references{
<https://squidle.org/api/help?template=api_help_page.html>
}
