% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{sq_instance}
\alias{sq_instance}
\title{Get or set the Squidle+ instance to use}
\usage{
sq_instance(url)
}
\arguments{
\item{url}{string: the URL of the instance to use. By default `squidleplus` uses "https://squidle.org"}
}
\value{
If called with no arguments, the current instance URL will be returned as a string. If called with a \code{url} argument, the instance will be set to that and then the instance URL returned.
}
\description{
Get or set the Squidle+ instance to use
}
\examples{

## current instance
sq_instance()

## set new instance
sq_instance("https://my.squidleinstance.com")

}
