% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{sq_query}
\alias{sq_query}
\title{Construct a request object}
\usage{
sq_query(endpoint, query, results_per_page = 10, page = 1)
}
\arguments{
\item{endpoint}{string: e.g. `/api/media`}

\item{query}{list: as returned by [sq_req_query()]}

\item{results_per_page}{integer: results are paginated - how many per page?}

\item{page}{integer: results are paginated - which page to return?}
}
\value{
An `httr2_request` object
}
\description{
Construct a request object
}
\examples{
\dontrun{
  ## search for all media that contains annotations with the name that
  ##  includes "ecklonia" (ignoring case)
  res <- sq_query("api/media", query = sq_req_query(filters = sq_v(annotations \%any\%
                 sq_v(annotations \%any\% sq_v(label \%has\% sq_v(name \%ilike\% "\%ecklonia\%"))))))
  sq_req_content(res)
}
}
