% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{sq_req_content}
\alias{sq_req_content}
\title{Extract response content}
\usage{
sq_req_content(res)
}
\arguments{
\item{res}{httr2_response:}
}
\value{
An R object
}
\description{
Extract response content
}
