% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{to_json}
\alias{to_json}
\alias{from_json}
\title{Convert an R object to a JSON string and vice-versa}
\usage{
to_json(x)

from_json(x)
}
\arguments{
\item{x}{object: thing to convert}
}
\value{
For `to_json` a string, converted to JSON using `jsonlite::toJSON(x, auto_unbox = TRUE)`. For `from_json` an R object, converted using `jsonlite::fromJSON(x)`
}
\description{
Convert an R object to a JSON string and vice-versa
}
\seealso{
[jsonlite::toJSON()] [jsonlite::fromJSON()]
}
